Initialize event detectors in DataPointEventNotifyWorkItem instead of calling thread
Intialize event detectors in DataPointEventNotifyWorkItem
