+++ Core/src/com/serotonin/m2m2/rt/dataImage/DataPointRT.java
@@ -25,6 +25,7 @@ import com.serotonin.m2m2.DataTypes;
 import com.serotonin.m2m2.db.dao.PointValueCacheDao;
 import com.serotonin.m2m2.db.dao.PointValueDao;
 import com.serotonin.m2m2.db.dao.SystemSettingsDao;
+import com.serotonin.m2m2.rt.DataPointEventNotifyWorkItem;
 import com.serotonin.m2m2.rt.dataImage.types.BinaryValue;
 import com.serotonin.m2m2.rt.dataImage.types.DataValue;
 import com.serotonin.m2m2.rt.dataImage.types.ImageValue;
@@ -32,10 +33,8 @@ import com.serotonin.m2m2.rt.dataImage.types.MultistateValue;
 import com.serotonin.m2m2.rt.dataImage.types.NumericValue;
 import com.serotonin.m2m2.rt.dataSource.PointLocatorRT;
 import com.serotonin.m2m2.rt.event.detectors.PointEventDetectorRT;
-import com.serotonin.m2m2.rt.maint.work.WorkItem;
 import com.serotonin.m2m2.rt.script.AbstractPointWrapper;
 import com.serotonin.m2m2.rt.script.DataPointWrapper;
-import com.serotonin.m2m2.util.ExceptionListWrapper;
 import com.serotonin.m2m2.util.timeout.TimeoutClient;
 import com.serotonin.m2m2.util.timeout.TimeoutTask;
 import com.serotonin.m2m2.view.stats.IValueTime;
@@ -46,8 +45,6 @@ import com.serotonin.m2m2.vo.event.detector.AbstractPointEventDetectorVO;
 import com.serotonin.timer.AbstractTimer;
 import com.serotonin.timer.FixedRateTrigger;
 import com.serotonin.timer.OneTimeTrigger;
-import com.serotonin.timer.RejectedTaskReason;
-import com.serotonin.timer.Task;
 import com.serotonin.timer.TimerTask;
 import com.serotonin.util.ILifecycle;
 
@@ -760,97 +757,8 @@ public class DataPointRT implements IDataPointValueSource, ILifecycle {
             boolean backdate, boolean logged, boolean updated, boolean attributesChanged) {
         DataPointListener l = Common.runtimeManager.getDataPointListeners(vo.getId());
         if (l != null)
-            Common.backgroundProcessing.addWorkItem(new EventNotifyWorkItem(vo.getXid(), l, oldValue, newValue,
+            Common.backgroundProcessing.addWorkItem(new DataPointEventNotifyWorkItem(vo.getXid(), l, oldValue, newValue,
                     attributes, set, backdate, logged, updated, attributesChanged));
-    }
-
-    class EventNotifyWorkItem implements WorkItem {
-        private static final String descriptionPrefix = "Point event for: ";
-        private static final String prefix = "EN-";
-        private final String sourceXid;
-        private final DataPointListener listener;
-        private final PointValueTime oldValue;
-        private final PointValueTime newValue;
-        private final Map<String, Object> attributes;
-        private final boolean set;
-        private final boolean backdate;
-        private final boolean logged;
-        private final boolean updated;
-        private final boolean attributesChanged;
-
-        EventNotifyWorkItem(String xid, DataPointListener listener, PointValueTime oldValue, PointValueTime newValue, Map<String, Object> attributes, boolean set,
-                boolean backdate, boolean logged, boolean updated, boolean attributesChanged) {
-            this.sourceXid = xid;
-            this.listener = listener;
-            this.oldValue = oldValue;
-            this.newValue = newValue;
-            this.attributes = attributes;
-            this.set = set;
-            this.backdate = backdate;
-            this.logged = logged;
-            this.updated = updated;
-            this.attributesChanged = attributesChanged;
-        }
-
-        @Override
-        public void execute() {
-            try {
-                if (attributesChanged) {
-                    listener.attributeChanged(attributes);
-                    return;
-                }
-
-                if (backdate)
-                    listener.pointBackdated(newValue);
-                else if (updated) {
-                    // Updated
-                    listener.pointUpdated(newValue);
-
-                    // Fire if the point has changed.
-                    if (!PointValueTime.equalValues(oldValue, newValue))
-                        listener.pointChanged(oldValue, newValue);
-
-                    // Fire if the point was set.
-                    if (set)
-                        listener.pointSet(oldValue, newValue);
-                }
-
-                // Was this value actually logged
-                if (logged)
-                    listener.pointLogged(newValue);
-            } catch(ExceptionListWrapper e) {
-                LOG.warn("Exceptions in event notify work item.");
-                for(Exception e2 : e.getExceptions())
-                    LOG.warn("Listener exception: " + e2.getMessage(), e2);
-            }
-        }
-
-        @Override
-        public int getPriority() {
-            return WorkItem.PRIORITY_MEDIUM;
-        }
-
-        @Override
-        public String getDescription() {
-            return descriptionPrefix + sourceXid;
-        }
-
-        @Override
-        public String getTaskId() {
-            //So there is one task for each listener
-            return prefix + sourceXid + "-" + listener.hashCode();
-        }
-
-        @Override
-        public int getQueueSize() {
-            return Task.UNLIMITED_QUEUE_SIZE;
-        }
-
-        @Override
-        public void rejected(RejectedTaskReason reason) {
-            //No special handling, tracking/logging is handled by the WorkItemRunnable
-        }
-
     }
 
     //
